DepthController: Fix blur depth radius inconsistensies
* limit to fwb's maximum blur radius
* dp -> px (getBlurRadius actually returns px instead of dp due to the pref value being int and wasnt being passed on getDimension function)
* reset the default value to report integer instead of dp.
* set minimum to 1px

+++ quickstep/src/com/android/launcher3/statehandlers/DepthController.java
@@ -126,7 +126,7 @@ public class DepthController implements StateHandler<LauncherState>,
     /**
      * Blur radius when completely zoomed out, in pixels.
      */
-    private float mMaxBlurRadius;
+    private int mMaxBlurRadius;
     private boolean mCrossWindowBlursEnabled;
     private WallpaperManager mWallpaperManager;
     private SurfaceControl mSurface;
